+++ feature/src/main/java/com/teamkkumul/feature/utils/time/GetCurrentTime.kt
@@ -1,4 +1,4 @@
-package com.teamkkumul.feature.utils
+package com.teamkkumul.feature.utils.time
 
 import java.text.SimpleDateFormat
 import java.time.LocalDateTime
